feat(cli): implement core CLI commands for package management and download functionality
