+++ internal/cli/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"bytes"
@@ -29,7 +29,7 @@ func TestDefaultData(t *testing.T) {
 func TestRun_Version(t *testing.T) {
 	Version = "1.2.3"
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"ver"}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"ver"}, &stdout, &stderr, nil)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -43,7 +43,7 @@ func TestRun_Version(t *testing.T) {
 
 func TestRun_RequireSubcommand(t *testing.T) {
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{}, &stdout, &stderr, nil)
+	exitCode := Run([]string{}, &stdout, &stderr, nil)
 	if exitCode != 1 {
 		t.Fatalf("expected exit code 1, got %d", exitCode)
 	}
@@ -54,7 +54,7 @@ func TestRun_RequireSubcommand(t *testing.T) {
 
 func TestRun_Help(t *testing.T) {
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"help"}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"help"}, &stdout, &stderr, nil)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -68,7 +68,7 @@ func TestRun_Help(t *testing.T) {
 
 func TestRunPkg_RequireSubcommand(t *testing.T) {
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"pkg"}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"pkg"}, &stdout, &stderr, nil)
 	if exitCode != 1 {
 		t.Fatalf("expected exit code 1, got %d", exitCode)
 	}
@@ -79,7 +79,7 @@ func TestRunPkg_RequireSubcommand(t *testing.T) {
 
 func TestRunPkgAdd_RequireArgument(t *testing.T) {
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"pkg", "add"}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"pkg", "add"}, &stdout, &stderr, nil)
 	if exitCode != 1 {
 		t.Fatalf("expected exit code 1, got %d", exitCode)
 	}
@@ -90,7 +90,7 @@ func TestRunPkgAdd_RequireArgument(t *testing.T) {
 
 func TestRun_RequireManifestArgument(t *testing.T) {
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"dl"}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"dl"}, &stdout, &stderr, nil)
 	if exitCode != 1 {
 		t.Fatalf("expected exit code 1, got %d", exitCode)
 	}
@@ -101,7 +101,7 @@ func TestRun_RequireManifestArgument(t *testing.T) {
 
 func TestRun_PathNotFound(t *testing.T) {
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"dl", "missing.yml"}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"dl", "missing.yml"}, &stdout, &stderr, nil)
 	if exitCode != 2 {
 		t.Fatalf("expected exit code 2, got %d", exitCode)
 	}
@@ -117,7 +117,7 @@ func TestRun_ParseError(t *testing.T) {
 		t.Fatalf("failed to write bad file: %v", err)
 	}
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"dl", badFile}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"dl", badFile}, &stdout, &stderr, nil)
 	if exitCode != 3 {
 		t.Fatalf("expected exit code 3, got %d", exitCode)
 	}
@@ -134,7 +134,7 @@ func TestRun_Spider(t *testing.T) {
 		t.Fatalf("failed to write yaml: %v", err)
 	}
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"dl", "--spider", yamlPath}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"dl", "--spider", yamlPath}, &stdout, &stderr, nil)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -168,7 +168,7 @@ func TestRun_DownloadSuccess(t *testing.T) {
 		return 123, nil
 	}
 
-	exitCode := run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
+	exitCode := Run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -196,7 +196,7 @@ func TestRun_DownloadAbsoluteRename(t *testing.T) {
 		return 0, nil
 	}
 
-	exitCode := run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
+	exitCode := Run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -228,7 +228,7 @@ func TestRun_RemoteYAML(t *testing.T) {
 		return 1, nil
 	}
 
-	exitCode := run([]string{"dl", server.URL + "/config.yml"}, &stdout, &stderr, downloader)
+	exitCode := Run([]string{"dl", server.URL + "/config.yml"}, &stdout, &stderr, downloader)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -254,7 +254,7 @@ func TestRun_DownloadError(t *testing.T) {
 	}
 
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
+	exitCode := Run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
 	if exitCode != 4 {
 		t.Fatalf("expected exit code 4, got %d", exitCode)
 	}
@@ -290,7 +290,7 @@ func TestRun_DownloadDigestMatch(t *testing.T) {
 		return int64(len(fileContent)), nil
 	}
 
-	exitCode := run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
+	exitCode := Run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -322,7 +322,7 @@ func TestRun_DownloadDigestMismatch(t *testing.T) {
 		return int64(len(fileContent)), nil
 	}
 
-	exitCode := run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
+	exitCode := Run([]string{"dl", yamlPath}, &stdout, &stderr, downloader)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d", exitCode)
 	}
@@ -343,7 +343,7 @@ func TestRunPkgAdd_Success(t *testing.T) {
 	}
 
 	var stdout, stderr bytes.Buffer
-	exitCode := run([]string{"pkg", "add", manifest}, &stdout, &stderr, nil)
+	exitCode := Run([]string{"pkg", "add", manifest}, &stdout, &stderr, nil)
 	if exitCode != 0 {
 		t.Fatalf("expected exit code 0, got %d (stderr=%s)", exitCode, stderr.String())
 	}
